Add python level function for xcorr1.
xcorr1 is a function to compute one-dimensional cross-correlation for
arbitrary axis and arbitrary lags. It can be useful when only a couple
of lags are necessary, in which case it is likely to be faster. If most
lags of the cross-correlation are necessary, FFT-based correlations are
of course much faster.